requirements.txt
aniso8601==8.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-RESTful==0.3.8
get==2019.4.13
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==7.0.0
post==2019.4.13
public==2019.4.13
PySocks==1.7.1
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
tweepy==3.8.0
urllib3==1.25.8
Werkzeug==1.0.0
